Test helpers pass all changes up (#499)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-4

While running through our new test helpers we noticed one small issue; although changes cascaded up to the next level (eg. from transaction to licence), they did not propagate any further than that. We quickly realised that although our add methods would call the update method of the parent, the update methods were not calling their parents. This change resolves this.


* Rename helper parameters

We realised that the `entity` parameter in each helper's `update` method was not consistent with the `add` method, where we use the actual type of entity (eg. `billRun`). We therefore renamed this across the helpers.

* `update` method of `NewInvoiceHelper` updates bill run

We amend the `update` method so that when we call it with updates (ie. numbers to be added to the counts/values) we also call the `update` method of the parent bill run with those updates

* `update` method of `NewLicenceHelper` updates invoice

* Update `NewTransactionHelper` tests

Although we're confident that the other helpers are working correctly, we add additional tests to `NewTransactionHelper` to check that the bill run and invoice are updated, bringing it in line with our other helpers' revised tests (which check that all levels are updated)