Error handling
There are two possible failure scenarios:

The first is where exporting tables to files fails. In this case, we wish to log an error using notifier and abort the process.

The second is where sending files to S3 fails. In this case, we wish to log an error and continue sending files.

This commit adds try/catch blocks to the export and send methods to handle these cases accordingly.

We also update the service to return `true` if all tables were sent successfully, and `false` if they weren't (either due to exporting failing, or at least one failing to send). This can be used by the admin endpoint to gauge success.