Create `DataExportService` (#501)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-67

Now that we have created TableExportService, we turn our attention to DataExportService which calls TableExportService to export a set of db tables as CSV files, and uses SendFileToS3Service to put them in the top-level csv 'folder' of an S3 bucket (inverted commas around 'folder' to acknowledge that S3 strictly speaking doesn't have folders, just key names that look like folders, but tools generally interpret them as a folder structure).

The service has error handling to satisfy the ACs:

If exporting a table fails then an error is logged using notifier and the entire process is halted;
If sending a file to the bucket fails then an error is logged using notifier and the process continues.
The service also returns a boolean to indicate whether the process succeeded:

If all tables are successfully exported and sent then it returns true;
If a table fails to export then it returns false;
If all tables are successfully exported, but one or more tables fail to send then it returns false.


* Create` ReportingExportService`

https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-67

Now that we have created `TableExportService`, we turn our attention to `ReportingExportService` which calls `TableExportService` to export a set of db tables, and uses `SendFileToS3Service` to put them in an S3 bucket.

* Initial outline of service

* Change `ExportTableService` file extension to `csv`

We realised that because our exported files will eventually be read in as CSV by a reporting tool, we would want the file extension to be `.csv`.

* Fix test stub extension

When we updated `ExportTableService` to produce files with a `.csv` extension, we neglected to update the stub in the `ReportingExportService` unit test to return the appropriate extension.

* Rename service

We initially named this `ReportingExportService`. After discussion we have renamed it to `DataExportService`.

* Send files to correct location

First we fix a small test error which was checking the wrong argument. Then we ensure the remote path starts with `csv`. Finally we amend the `SendFileToS3Service` call to ensure files are not sent to the archive bucket.

* Remove stray .only

* Error handling

There are two possible failure scenarios:

The first is where exporting tables to files fails. In this case, we wish to log an error using notifier and abort the process.

The second is where sending files to S3 fails. In this case, we wish to log an error and continue sending files.

This commit adds try/catch blocks to the export and send methods to handle these cases accordingly.

We also update the service to return `true` if all tables were sent successfully, and `false` if they weren't (either due to exporting failing, or at least one failing to send). This can be used by the admin endpoint to gauge success.

* Update comments

* Update app/services/data_export.service.js

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>