Fix proxy config defect (#511)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-98

As reported by WRLS, the existing way of specifying the proxy doesn't work correctly. This is due to us replacing the previous http request library with Got but not updating the proxy configuration to reflect breaking changes, and this went undiscovered as we don't use a proxy in any of our environments! We therefore resolve this by reworking how proxy configuration is applied.

In order to test this locally, add the following lines to docker-compose.yml:

  proxy:
    image: hinata/nginx-forward-proxy
    ports:
      - "3128:3128"
And define the proxy in .env as follows:

HTTP_PROXY=http://proxy:3128


* Rework proxy configuration

* Remove unit tests

As this is a straightforward fix to how we were passing through configuration options, we decided we don't need to add additional unit tests to our existing rules service tests

* Info on enabling local proxy

So that we don't lose the info on how to set up a local proxy for testing, we add comments to docker-compose.yml and .env.example detailing the process.