Move controllers from presroc folder (#513)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-101

As part of our preparations for the forthcoming sroc work, the dev team had discussions on how we should be organising our files and arrived at the conclusion that the presroc folder our controllers lived in was no longer necessary, and that we would be adopting a different pattern for organising controllers in future. We therefore need to move our controllers out of app/controllers/presroc and into app/controllers.

While doing this, we also take the opportunity to rename the controllers to remove Presroc from the name (ie. PresrocBillRunsController becomes simply BillRunsController).


* Move controllers out of presroc folder

* Remove `Presroc` from controller names

* Update test descriptions