[CMEA-102] Move services into folders (#514)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-102

In order to clean up our services folder, we create folders and organise the services into them. Each service is moved into a folder according to the entity it relates to. So bill run-related services are in the bill_runs folder etc. We have also taken the opportunity to rename some of the services to make their intent clearer. Tests have been structured accordingly as well.

Note that this PR fails the SonarCloud code analysis. This is due to moving files around causing previously-ignored code smells to be flagged up once again. On inspection we can resolve some of these easily; some of them will take a little longer to resolve; and some can be ignored entirely (eg. we don't propose replacing a string with a constant when this would make the code harder to read). However in the interests of keeping this PR entirely focused on the restructuring of our services, we will tackle these in a separate PR.


* [CMEA-102] Move services into function folders

https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-102

In order to clean up our `services` folder, we create function folders and organise the services into them.

* Initial organisation

The first stage is to move the services into their new folders. We expect the service and all unit tests to fail at this stage! Note that as we went, some references were automatically updated by VS Code and some weren't. We also renamed some services in the process.

* Service starts

This updates all requires to allow the service to start.

* Fix broken unit test

* Rename services

* Rename tests

* Move test files into folders

* Sort `index.js`