Delete `REMOVE_TMP_FILES` environment variable (#517)
One of the SonarCloud warnings we reviewed related to the REMOVE_TMP_FILES environment variable (which controls whether customer and transaction files should be deleted after being sent to S3. We came to the conclusion that we don't actually need this environment variable -- it defaults to true, isn't currently set in any of our environments, and the chances of us ever needing to set it are exceedingly slim. We are therefore removing it.

* Delete `removeTemporaryFiles` from `ServerConfig`

* Update `SendTransactionFileService`

* Update `SendCustomerFileService`

* Update service comments

* Remove environment variable from `.env.example`

* Update unit tests