Add Docker build and push to GitHub workflows (#522)
Back at the start of the year we linked our repo to [Docker Hub](https://hub.docker.com/repository/docker/environmentagency/sroc-charging-module-api) and enabled auto-bulding of images. This meant as soon as changes were made to `main` or a new tag was pushed Docker Hub would automatically build the image for us. The reasoning behind this was to make it possible for client teams (currently the @DEFRA/water-abstraction team) to run the API locally rather than having to rely on our hosted AWS environments.

That worked beautifully and we thought no more about it. That was until dropping into Docker Hub last week we spotted the last 2 tags did not have corresponding images. A further check showed **latest** had not been updated in months. Some frantic digging tracked down this post

> Unfortunately, Docker’s Autobuild service has been targeted by the same bad actors, so today we are disappointed to announce that we will be discontinuing Autobuilds on the free tier starting from June 18, 2021.
>
> https://www.docker.com/blog/changes-to-docker-hub-autobuilds/

So, that ability to run the CM locally has been broken. As an alternate solution, we've researched and found we can use GitHub actions to replicate what Docker Hub's autobuild did. At the same time, we found some other new restrictions to Docker Hub for the free tier. Use of more than one access token has been put behind the paywall as well as access to the better UI (BTW we understand why they need to make these changes).

There is still some internal debate going on about which registry teams should be using but these changes have also caused us to look elsewhere. Turns out [GitHub has a container registry!](https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-container-registry)

So, to begin with we're going to build and push our images there. The building of the images won't change but we may push them somewhere else in the future. For now, this change adds a new Docker build and push workflow that will be triggered

- whenever a change is made to `main`
- whenever a tag with a [semver pattern](https://semver.org/) is pushed