Add pushing to DockerHub to Docker workflow
In [Add Docker build and push to GitHub workflows](https://github.com/DEFRA/sroc-charging-module-api/pull/522) we added a new workflow to automatically build Docker images after pushes to `main` and when new tags are added. We also settled on pushing to [GitHub Container registry](https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-container-registry) as our primary registry.

But we also highlighted that there is still some discussion about whether we should be using a single registry (and which one) plus our existing users are pulling things from [DockerHub](https://hub.docker.com/repository/docker/environmentagency/sroc-charging-module-api). So, to continue to support our clients and in case DockerHub wins out, this change amends the Docker workflow to also push to DockerHub.