Add custom sections to CHANGELOG (#525)
We have always used the tool [github-changelog-generator](https://github.com/github-changelog-generator/github-changelog-generator) to generate our `CHANGELOG.md`. It does this based on your PR history which is why we also try and ensure meaningful titles and descriptions are used.

We also use some non-default [labels](https://docs.github.com/en/issues/using-labels-and-milestones-to-track-work/managing-labels) for our PR's and have always wondered how we could include them in the changelog. Anything with a custom label has always gone under the catch-all **Merged PR's** section.

Today we finally found the answer for [how to tell it to add sections](https://github.com/github-changelog-generator/github-changelog-generator/wiki/Advanced-change-log-generation-examples#additional-options) based on labels you've added.

This change updates the `CHANGELOG.md` by regenerating it with this extra config which causes the new sections to appear. It should be noted no content is dropped. It is just more clearly grouped.

It also adds a config file that contains the JSON used to tell github_changelog_generator what sections to add and what labels to link them with. We don't really want to have to try and remember to add this each time!

By adding this [Params file](https://github.com/github-changelog-generator/github-changelog-generator#params-file) github_changelog_generator will automatically use it if the command `github_changelog_generator` is run from the root of the project.