Add dev guide amending code to handle new rulesets
https://eaflood.atlassian.net/browse/CMEA-103

Currently, the CHA only supports bill runs and transaction information for PRE-SROC charges. We are expecting in the coming months to start updating the service to handle [SROC](https://www.nfuonline.com/cross-sector/environment/water/irrigation-and-water-resources/ea-launches-consultation-on-strategic-review-of-abstraction-charges/) based charges.

The financial year a charge is applicable to determines the ruleset used. In the context of the CHA this means we'll be calling a different Rules Service endpoint and might mean we need to validate data differently or amend the behaviour of a service. We've done some initial work to prepare for this (See PRs #513, #514, and #519) and even done a [Sroc dev plan spike](https://github.com/DEFRA/sroc-charging-module-api/pull/523).

Thanks to this we can now document how to amend the project when its been identified a change is needed to support differences in the rulesets.