Add pushing to DockerHub to Docker workflow (#527)
In [Add Docker build and push to GitHub workflows](https://github.com/DEFRA/sroc-charging-module-api/pull/522) we added a new workflow to automatically build Docker images after pushes to `main` and when new tags are added. We also settled on pushing to [GitHub Container registry](https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-container-registry) as our primary registry.

But we also highlighted that there is still some discussion about whether we should be using a single registry (and which one) plus our existing users are pulling things from [DockerHub](https://hub.docker.com/repository/docker/environmentagency/sroc-charging-module-api). So, to continue to support our clients and in case DockerHub wins out, this change amends the Docker workflow to also push to DockerHub.

** Notes about the change

We have to log in to any registries we interact with and having initially played with [Tag-Push action](https://github.com/marketplace/actions/tag-push-action) and then following the docs for multi-registries with [Build-Push action](https://github.com/docker/build-push-action/blob/master/docs/advanced/push-multi-registries.md) consensus seems to be login to your main registry last.

We are depending on [Build-Push action's](https://github.com/marketplace/actions/build-and-push-docker-images) ability to push to multiple registries but that means we need to give it all our tags in one go.

So, we have another new step where we combine what we were already getting from [Metadata action](https://github.com/marketplace/actions/docker-metadata-action) with the output of a [sed](https://en.wikipedia.org/wiki/Sed) call (which replaces `ghcr.io` with `docker.io`).

It requires a bit of extra work to deal with multiline strings because if, for example, a new tag is pushed Metadata action will return 2 tags (`v*.*.*` and `latest`) split over 2 lines. With our sed call we then end up with a 4 line string we need to hold in an env var.

Once we've got the env var though, the final change is to refer to that rather than `steps.meta.outputs.tags` in our 'Build and push Docker image' step.

The end result is whether the push is a change to `main` or a new tag, whatever Docker tags we push to GitHub Container registry will also get pushed to DockerHub!