Implement initial fix
Currently, each eleent going into the CSV file is simply put in with quotes around it. This works for everything except JSON objects (ie. `transactions.line_description`), where this gives us `"[object Object]"`.

As an initial attempt at fixing this, we add some additional parsing to `StreamTransformCSVService`. The first step is to run JSON.stringify() on any objects. The second step (which we know we need to do after looking at what TablePlus outputs when exporting a table to CSV) is to replace all quotes " with two quotes "".

This works for JSON objects, but dates pulled from the db also have `typeof` `'object'`. Stringifying them causes two issues:

The first is that it appears to put quotes around them, so when we replace " with "", we end up with two quotes at the start and end of the string.

The second is that it puts them into a different date format than when we simply put quotes around them. Previously we had dates in the format `Fri Sep 03 2021 13:38:22 GMT+0000 (Coordinated Universal Time)`, we now have dates in the format `2021-09-03T13:38:22.769Z`. This is arguably easier to parse, however it slightly breaks our `TransformRecordsToFileService` tests, which expect dates in the previous format.