Fix CSV transform object defect (#534)
https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-105

When we export a table to CSV, we use our StreamTransformCSVService which performs a simple transformation to convert the data to CSV format. This works for all data types with the exception of data stored in the db in JSON which converts to [object Object].

This branch fixes this by checking whether the conversion would result in [object Object]. If it doesn't then we don't need to worry about it. If it does then we use JSON.stringify() to give us the object in text form. Note that we don't simply use typeof to determine the element's type, as it also identifies dates as objects and stringifying them changes the format giving us a less-than ideal result.

We also add an additional stage, which checks whether the element is a string and if it is, sanitises any quotes by replacing each one with two quotes (ie. from this is a "quote" to this is a ""quote""). We realised we needed this after exporting a table to CSV in TablePlus, and observing the format of the output.

* Fix CSV transform object defect

* Create unit test

* Additional test scenario

* Rework test

* Implement initial fix

Currently, each eleent going into the CSV file is simply put in with quotes around it. This works for everything except JSON objects (ie. `transactions.line_description`), where this gives us `"[object Object]"`.

As an initial attempt at fixing this, we add some additional parsing to `StreamTransformCSVService`. The first step is to run JSON.stringify() on any objects. The second step (which we know we need to do after looking at what TablePlus outputs when exporting a table to CSV) is to replace all quotes " with two quotes "".

This works for JSON objects, but dates pulled from the db also have `typeof` `'object'`. Stringifying them causes two issues:

The first is that it appears to put quotes around them, so when we replace " with "", we end up with two quotes at the start and end of the string.

The second is that it puts them into a different date format than when we simply put quotes around them. Previously we had dates in the format `Fri Sep 03 2021 13:38:22 GMT+0000 (Coordinated Universal Time)`, we now have dates in the format `2021-09-03T13:38:22.769Z`. This is arguably easier to parse, however it slightly breaks our `TransformRecordsToFileService` tests, which expect dates in the previous format.

* Revised fix

We realised that if we only want to stringify objects that result in `[object Object]` when wrapped in quotes, we just to need to check for thia -- if it does then we stringify it, otherwise we don't bother.

* Revised revised fix

We somehow managed to break the fix just prior to commiting it by deleting the part where we assign the output of `element.replace()`