Create separate CSV and dat transform services
We realised that the format we were calling "CSV" was not really CSV format and not well-suited for importing due to every element being wrapped in quotes so numbers, dates etc. may be misidentified as strings. Since the customer and transaction files (which use the "everything must be wrapped in quotes" format) are .dat files, we will refer to this as "dat" format. We therefore create a separate `StreamTransformDatRowService` which will wrap everything in quotes, allowing our `StreamTransformCSVRowService` to correctly handle CSV formatting.