Amend CSV export format (#537)
While fixing the CSV export object defect, we found that we could easily turn the date output format into a more friendly format by using JSON.stringify() on the date object. This has turned into a wider change to correctly output CSV-formatted data; for example, with quotes around strings but not around integers. This is distinct from the dat format used for customer and transaction files, which has quotes around everything.

We start by renaming StreamTransformCSVService to StreamTransformCSVRowService to be consistent with the new StreamTransformDatRowService.

We then refactor the logic out of StreamTransformCSVRowService and into a separate ConvertToCSVService.

Finally, we update this service to correctly handle different tapes of data.


* Refactor code into `ConvertToCSVService`

* Rename `StreamWritableFileService` test file

We noticed that the test file for `StreamWritableFileService` didn't have `.test` in the filename.

* Divide up unit tests

After refactoring, we move the output tests out of the `StreamTransformCSVService` unit tests and into `ConvertToCSVService` tests

* Improved object handling

We amend object formatting to stringify all objects, which will give us nicer date output. Doing this will wrap the date string in quotes, so we ensure they're removed prior to the stage of replacing " with "".

* Fix `ExportTableToFileService` test

Our changes break the `ExportTableToFileService` tests, which expect the date in a specific format. We therefore amend the tests to expect the new format.

* Document service

* Create separate CSV and dat transform services

We realised that the format we were calling "CSV" was not really CSV format and not well-suited for importing due to every element being wrapped in quotes so numbers, dates etc. may be misidentified as strings. Since the customer and transaction files (which use the "everything must be wrapped in quotes" format) are .dat files, we will refer to this as "dat" format. We therefore create a separate `StreamTransformDatRowService` which will wrap everything in quotes, allowing our `StreamTransformCSVRowService` to correctly handle CSV formatting.

* Correct handling of non-strings

* Update app/services/streams/stream_transform_csv_row.service.js

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>

* Move `formatDate()` to `GeneralHelper`

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>